Refactor analysis classes and add winsorizing data to calculate better mean and stddev
