+++ pyproject.toml
@@ -4,8 +4,9 @@
 #
 # pip install -e .
 #
-# to add it to the environment. Main reason for this is to make it easier for
-# people who just want to use mainly the segmentation and not develop it.
+# to add it to the environment.
+# The main reason for this is to make it easier for
+# people who just want to use the segmentation and not develop it.
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
@@ -27,7 +28,7 @@ packages = [
 ]
 
 [tool.poetry.scripts]
-synthSeg-mpm-analyse = "SynthSeg.analysis.label_analysis:main"
+synthSeg-mpm-analyse = "SynthSeg.analysis.mpm_analysis:main"
 synthSeg-cbs-predict = "scripts.commands.cbs_predict:main"
 synthSeg-predict = "scripts.commands.SynthSeg_predict:main"
 synthSeg-training = "scripts.commands.training:main"
